fix(apiClient): mount MigrationBlockModal inside ApiClientProvider
The modal sat as a sibling of ApiClientFeatureContainer in
RouteElement, which placed it outside ApiClientProvider — the wrapper
that supplies the per-workspace context the export hooks
(useAllRecords, useAllEnvironments, useApiClientFeatureContext) need.
Result: runtime "could not find react-redux context value" error when
useWorkspaceZipDownload tried to call useAllRecords from inside the
modal.

Move the show-flag + segment check + modal mount into container.tsx,
inside <ApiClientProvider>. RouteElement returns to its pre-RQ-1806
shape.
